Split sample app repo from diffing code
